requirements.txt
lxml>=3.1.0
typing_extensions>=4.9.0
python-docx>=1.2.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
docxtpl>=0.16.0
